Add methods for automatically detecting the Pi-Hole
As soon as the user starts up the app, we'll first try to ping their DNS
server, as that should in theory be the Pi-Hole. If that doesn't work,
then we offer the user the chance to either scan their network or
manually enter in the IP address for their Pi.