Fix scanning
the scanning functionality in and of itself worked just fine, but upon
successful connection, it would return to the scan or direct connection
screen, then go to the authentication screen, and then return to the
previous screen. This little hack prevents that but it should really be
replaced with a better solution.